move OrcaPoolToken to types.ts
