+++ src/model/orca/pool/pool-types.ts
@@ -1,5 +1,5 @@
 import { PublicKey } from "@solana/web3.js";
-import { OrcaToken } from "../../..";
+import { OrcaPoolToken } from "../../..";
 import { Percentage } from "../../../public";
 
 /**
@@ -26,14 +26,6 @@ export type OrcaPoolParams = {
   tokens: Record<string, OrcaPoolToken>;
   curveType: CurveType;
   feeStructure: FeeStructure;
-};
-
-/**
- * An Orca Token within an OrcaPool
- * @param addr The public key for this token for this Orca Pool
- */
-export type OrcaPoolToken = OrcaToken & {
-  addr: PublicKey;
 };
 
 export enum CurveType {
